refactor(records): the account roll-up is its own file
strength.go crossed the 500-line ceiling. The seam is not arbitrary:
one person's strength is a fact about that person, while an account's is
a fold over the contacts a caller may see — so it turns on the employment
EDGE, and the edge grant is asked in the account path and nowhere in the
per-person one. Two questions, two gates, two files.

Refs #1846
